support mmap returning address > 2^31
opf_anon_mmap would abort if mmap returned an address > 2^31 due to
sloppy coding checking for the return value of mmap being < 0.  That
worked fine on the 8MB machine running Linux 1.x that opf was original
written on and still worked fine on a 1GB machine running Linux 2.x.
However sometime in Linux 3.x mmap started returning addresses > 2^31
and so these would be flagged as an error.

Changed to check against 0xfffff000.  Only values above that are
considered to be an error.