Drop l suffix from most instructions
In most cases the assembler is smart enough to work out where the
instruction applies to a 32-bit value and so the explicit 'l' suffix
on the instruction is not needed.  Hopefully older versions of the
assembler were equally smart otherwise I've just broken building this
under an ancient version of Linux.